Stop pretending to use require (bug 1113801)
The old require() stub was very node-like which
means it is not compatible with RequireJS and this
caused problems when trying to plug fxpay into
Fireplace.

This patch isn't ideal but maybe we can fix it later
by doing proper module-fi-cation. All the patch does
is stop making namespace checks look like calls to
require() since they're not.